♻️ Move functions from _utils.py to _client.py
The following functions are moved because they are only used in _client.py
* port_or_default
* same_origin
* is_https_redirect

Related tests are also moved accordingly.
